ENH Move check for inferred compression to before `get_filepath_or_buffer`
When reading CSVs, if `compression='infer'`, check the input before calling `get_filepath_or_buffer` in the `_read` function. This way we can catch compresion extensions on S3 files.

We now attempt to infer compression from an input filename only in the `_read` function, instead of separately in each parser.