BUG: Fixed grow_buffer to grow when capacity is reached
Addresses issue in #12494 by allowing `grow_buffer` to grow the size
of the parser buffer when buffer capacity is achieved.  Previously,
you had to exceed capacity for this to occur, but that was
inconsistent with the `end_field` check later on when handling the EOF
terminator, where reached capacity was considered a buffer overflow.

Author: gfyoung <gfyoung17@gmail.com>

Closes #12504 from gfyoung/read_csv_empty_header and squashes the following commits:

8ba3dd0 [gfyoung] BUG: Fixed grow_buffer to grow when capacity is reached