DOC: fix code-block ipython highlighting
I noticed some warnings in the travis doc build log in the sense of
"WARNING: Could not parse literal_block as "python". highlighting
skipped."   This was because some of the `.. code-block:: python`
directives contained code that could not be interpreted as python.
And, additionally, these code blocks were also not highlighted
correctly (the `In [1]` prompt were not recognized by the default
python highlighter).     Therefore, I changed those `.. code-block::
python` directives containing IPython prompt to `.. code-block::
ipython`

Author: Joris Van den Bossche <jorisvandenbossche@gmail.com>

Closes #12853 from jorisvandenbossche/fix-ipython-highlighting and squashes the following commits:

1e104cc [Joris Van den Bossche] DOC: fix code-block ipython highlighting