BUG: Parse NULL char as null value
Fixes bug in C parser in which the `NULL` character (`'\x00'`) was
being interpreted as a true line terminator, escape character, or
comment character because it was used to indicate that a user had not
specified these values. As a result, if the data contains this value,
it was being incorrectly parsed. It should be parsed as `NULL`.
Closes #14012.

Author: gfyoung <gfyoung17@gmail.com>

Closes #14019 from gfyoung/null-char-parse and squashes the following commits:

5d39744 [gfyoung] BUG: Parse NULL char as null value