BUG: Parse two date columns broken in read_csv with multiple headers
In `io/parsers/_try_convert_dates()` when selecting
columns based on a  column index from a set of columns with multi-
level names, the column  `name` was converted to a string.  This
appears to be a bug since the  `name` was a tuple before the
conversion.  This causes problems  downstream when there is an attempt
to use this name to lookup a  column, and that lookup fails because
the desired column is keyed from  the tuple, not its string
representation

closes #15376

Author: Stephen Rauch <stephen.rauch+github@gmail.com>

Closes #15378 from stephenrauch/fix_read_csv_merge_datetime and squashes the following commits:

030f5ec [Stephen Rauch] BUG: Parse two date columns broken in read_csv with multiple headers